Integrated CUDA colorizer, but need verify result

+++ gems/cuda/colorizer.cu.hpp
@@ -5,7 +5,9 @@
 namespace isaac {
 
 // Converts a 1-channel 32-bit floating point depth image to a 3-channel 8-bit HUE image.
-void ImageF32ToHUEImage(StridePointer<const float> image, StridePointer<unsigned char> result,
-                               float min_depth, float max_depth, unsigned int width, unsigned int height);
+void ImageF32ToHUEImage(StridePointer<const float> image,
+                        StridePointer<unsigned char> result,
+                        float min_depth, float max_depth,
+                        size_t width, size_t height);
 
 }  // namespace isaac
